feat(#2): automatically supply the PIN code to unlock the SIM card (#20)
* Add `SIM PIN` preference

* Add `PinStorage` to store PIN codes in the CE storage

* Add `AuthenticationPromptActivity` to unlock hardware-backed KeyStore

* Add SIM PIN encrypt/decrypt support

* fix(TimePicker): call `callChangeListener()` only if time value changed

* Make available clear PIN codes after reboot & when alarm goes off

* Handle `Resume-on-Reboot` (RoR) stage when `Full-Disk Encryption` (FDE) enabled

* Add user-space (banner/notification) error handling on PIN incorrect/corrupted

* Add `SimPinFeeder` to supply SIM PIN codes when alarm goes off

* Import translations

* Update README.md

* Update app images

+++ src/com/github/iusmac/sevensim/ui/scheduler/SchedulerViewModel.java
@@ -24,11 +24,14 @@ import androidx.lifecycle.ViewModelProvider;
 import com.github.iusmac.sevensim.DateTimeUtils;
 import com.github.iusmac.sevensim.Logger;
 import com.github.iusmac.sevensim.R;
+import com.github.iusmac.sevensim.Utils;
 import com.github.iusmac.sevensim.scheduler.DayOfWeek;
 import com.github.iusmac.sevensim.scheduler.DaysOfWeek;
 import com.github.iusmac.sevensim.scheduler.SubscriptionScheduleEntity;
 import com.github.iusmac.sevensim.scheduler.SubscriptionScheduler;
 import com.github.iusmac.sevensim.scheduler.SubscriptionSchedulerSummaryBuilder;
+import com.github.iusmac.sevensim.telephony.PinEntity;
+import com.github.iusmac.sevensim.telephony.PinStorage;
 import com.github.iusmac.sevensim.telephony.Subscriptions;
 
 import dagger.assisted.Assisted;
@@ -45,6 +48,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 
 public final class SchedulerViewModel extends ViewModel {
@@ -86,6 +90,12 @@ public final class SchedulerViewModel extends ViewModel {
     private final MutableLiveData<LocalTime> mMutableEndTime;
     private LiveData<CharSequence> mObservableEndTimeSummary;
 
+    private final MediatorLiveData<Optional<PinEntity>> mMediatorPinEntity =
+        new MediatorLiveData<>(Optional.empty());
+    private LiveData<CharSequence> mObservablePinPresenceSummary;
+    private final MutableLiveData<Boolean> mMutablePinTaskLock = new MutableLiveData<>(false);
+    private LiveData<Optional<PinErrorMessage>> mObseravblePinErrorMessage;
+
     @SuppressLint("StaticFieldLeak")
     private final Context mContext;
     private final Logger mLogger;
@@ -93,6 +103,7 @@ public final class SchedulerViewModel extends ViewModel {
     private final SubscriptionScheduler mSubscriptionScheduler;
     private final Subscriptions mSubscriptions;
     private final SubscriptionSchedulerSummaryBuilder mSubscriptionSchedulerSummaryBuilder;
+    private final PinStorage mPinStorage;
     private final int mSubscriptionId;
 
     @AssistedInject
@@ -102,6 +113,7 @@ public final class SchedulerViewModel extends ViewModel {
             final SubscriptionScheduler subscriptionScheduler,
             final Subscriptions subscriptions,
             final SubscriptionSchedulerSummaryBuilder subscriptionSchedulerSummaryBuilder,
+            final PinStorage pinStorage,
             final @Assisted int subscriptionId,
             final @Assisted Looper looper) {
 
@@ -111,6 +123,7 @@ public final class SchedulerViewModel extends ViewModel {
         mSubscriptionScheduler = subscriptionScheduler;
         mSubscriptions = subscriptions;
         mSubscriptionSchedulerSummaryBuilder = subscriptionSchedulerSummaryBuilder;
+        mPinStorage = pinStorage;
         mSubscriptionId = subscriptionId;
 
         mResources = mContext.getResources();
@@ -121,6 +134,8 @@ public final class SchedulerViewModel extends ViewModel {
         mMutableDaysOfWeek = new MutableLiveData<>(mDaysOfWeekFactory.create());
         mMutableStartTime = new MutableLiveData<>(mMutableStartSchedule.getValue().getTime());
         mMutableEndTime = new MutableLiveData<>(mMutableEndSchedule.getValue().getTime());
+        mMediatorPinEntity.addSource(mPinStorage.getObservablePin(mSubscriptionId), (pinEntity) ->
+                mMediatorPinEntity.setValue(pinEntity));
 
         // Fetch data from the database
         mHandler.post(() -> {
@@ -290,6 +305,50 @@ public final class SchedulerViewModel extends ViewModel {
     }
 
     /**
+     * @return An observable human-readable status indicating PIN presence.
+     */
+    LiveData<CharSequence> getPinPresenceSummary() {
+        if (mObservablePinPresenceSummary == null) {
+            mObservablePinPresenceSummary = Transformations.map(mMediatorPinEntity, (pinEntity) ->
+                    mResources.getString(pinEntity.isPresent() ? R.string.scheduler_pin_set_summary
+                        : R.string.scheduler_pin_unset_summary));
+        }
+        return mObservablePinPresenceSummary;
+    }
+
+    /**
+     * @return An observable lock state of the SIM PIN task.
+     */
+    LiveData<Boolean> getPinTaskLock() {
+        return mMutablePinTaskLock;
+    }
+
+    /**
+     * @return An observable containing a human-readable PIN error message, if any.
+     */
+    LiveData<Optional<PinErrorMessage>> getPinErrorMessage() {
+        if (mObseravblePinErrorMessage == null) {
+            mObseravblePinErrorMessage = Transformations.map(mMediatorPinEntity, (pinEntity) ->
+                    pinEntity.map((pin) -> {
+                        final String title, reason;
+                        if (pin.isCorrupted()) {
+                            title = mResources.getString(R.string.sim_pin_operation_failed);
+                            reason = mResources.getString(
+                                    R.string.scheduler_pin_banner_sim_pin_decryption_failed_reason);
+                            return new PinErrorMessage(title, reason);
+                        } else if (pin.isInvalid()) {
+                            title = mResources.getString(R.string.sim_unlock_failed);
+                            reason = mResources.getString(
+                                    R.string.scheduler_pin_banner_wrong_sim_pin_code_reason);
+                            return new PinErrorMessage(title, reason);
+                        }
+                        return null;
+                    }));
+        }
+        return mObseravblePinErrorMessage;
+    }
+
+    /**
      * @param enabled {@code true} if the scheduler is being enabled, otherwise {@code false}.
      */
     void handleOnEnabledStateChanged(final boolean enabled) {
@@ -326,6 +385,45 @@ public final class SchedulerViewModel extends ViewModel {
             which == TimeType.START_TIME ? mMutableStartTime : mMutableEndTime;
         mutableTime.setValue(time);
         persist();
+    }
+
+    /**
+     * @param pin The SIM PIN as string.
+     */
+    void handleOnPinChanged(final @NonNull String pin) {
+        mLogger.d("handleOnPinChanged().");
+
+        mHandler.post(() -> {
+            // Acquire lock till asynchronous request completes
+            mMutablePinTaskLock.postValue(true);
+
+            final PinEntity pinEntity = mMediatorPinEntity.getValue().orElseGet(() -> {
+                final PinEntity p = new PinEntity();
+                p.setSubscriptionId(mSubscriptionId);
+                return p;
+            });
+            pinEntity.setInvalid(false);
+            pinEntity.setClearPin(pin);
+            if (mPinStorage.encrypt(pinEntity)) {
+                mPinStorage.storePin(pinEntity);
+            } else {
+                Utils.makeToast(mContext, mResources.getString(R.string.sim_pin_operation_failed));
+                // Unconditionally propagate an empty instance since unencrypted PIN entities cannot
+                // be persisted on disk, thus the mediator will remain unchanged
+                mMediatorPinEntity.postValue(Optional.empty());
+            }
+
+            // In order to supply the SIM subscription PIN codes to the active SIM subscriptions
+            // found on the device when processing schedules at the stated time, we need to
+            // re-schedule using the list of all decrypted SIM PIN entities
+            final List<PinEntity> pinEntities = mPinStorage.getPinEntities();
+            pinEntities.forEach((pinEntity1) -> mPinStorage.decrypt(pinEntity1));
+            mSubscriptionScheduler.updateNextWeeklyRepeatScheduleProcessingIter(
+                    LocalDateTime.now().plusMinutes(1), pinEntities);
+
+            // Release lock
+            mMutablePinTaskLock.postValue(false);
+        });
     }
 
     /**
@@ -371,7 +469,13 @@ public final class SchedulerViewModel extends ViewModel {
             mMutableEndSchedule.setValue(defaultSchedule);
         }
 
-        mHandler.post(() -> mSubscriptionScheduler.deleteAll(schedulesToRemove));
+        mMediatorPinEntity.getValue().ifPresent((pin) -> mHandler.post(() ->
+                    mPinStorage.deletePin(pin)));
+
+        if (!schedulesToRemove.isEmpty()) {
+            mHandler.post(() -> mSubscriptionScheduler.deleteAll(schedulesToRemove));
+        }
+
         refreshNextUpcomingScheduleSummaryAsync();
     }
 
@@ -381,6 +485,21 @@ public final class SchedulerViewModel extends ViewModel {
     boolean schedulerExists() {
         return mMutableStartSchedule.getValue().getId() > 0L ||
             mMutableEndSchedule.getValue().getId() > 0L;
+    }
+
+    /**
+     * @return {@code true} if the SIM PIN code has been set, otherwise {@code false}.
+     */
+    boolean isPinPresent() {
+        return mMediatorPinEntity.getValue().isPresent();
+    }
+
+    /**
+     * @return {@code true} if we need to authenticate the user with their credentials for further
+     * crypto operations on the SIM PIN code.
+     */
+    boolean isAuthenticationRequired() {
+        return mPinStorage.isAuthenticationRequired();
     }
 
     /**
@@ -440,6 +559,16 @@ public final class SchedulerViewModel extends ViewModel {
         mHandler.removeCallbacksAndMessages(null);
     }
 
+    final class PinErrorMessage {
+        final String title;
+        final String reason;
+
+        private PinErrorMessage(final String title, final String reason) {
+            this.title = title;
+            this.reason = reason;
+        }
+    }
+
     private final class IntentReceiver extends BroadcastReceiver {
         @Override
         public void onReceive(final Context context, final Intent intent) {
@@ -454,6 +583,7 @@ public final class SchedulerViewModel extends ViewModel {
                     mMutableDaysOfWeek.postValue(mMutableDaysOfWeek.getValue());
                     mMutableStartTime.postValue(mMutableStartTime.getValue());
                     mMutableEndTime.postValue(mMutableEndTime.getValue());
+                    mMediatorPinEntity.postValue(mMediatorPinEntity.getValue());
                     // Refresh the next upcoming schedule summary locale-sensitive part
                     refreshNextUpcomingScheduleSummaryAsync();
                     break;
